Squashed commit of the following:
commit 96a7df0d743a3456bc2a1cdb5b956804794abcc2
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 17 13:34:07 2026 -0800

    type hint read_ methods

commit dcb143dc84bac310200f1fcc5c7f33971e3c1514
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 17 12:35:18 2026 -0800

    add props passback to news

commit 558a5f1f3b23f1b5824fa7ccc7235415856a98c1
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 17 12:32:23 2026 -0800

    update deps

commit 10821c86178528c6f6f0b10deafdb4ca1ea8ae11
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 17 12:32:15 2026 -0800

    remove old format transform

commit f697d789b0c3eba266039decfa689392e42bb588
Author: AJ Slater <aj@slater.net>
Date:   Mon Feb 16 23:38:23 2026 -0800

    extract pages too
