+++ pyproject.toml
@@ -18,7 +18,7 @@ readme = "README.md"
 license = "GPL-3.0-only"
 name = "comicbox-pdffile"
 requires-python = "<4.0,>=3.10"
-version = "0.3.0"
+version = "0.4.0"
 [[project.authors]]
 name = "AJ Slater"
 email = "aj@slater.net"
@@ -190,7 +190,7 @@ addopts = [
 ]
 junit_family = "xunit2"
 strict = true
-testpaths = "tests"
+testpaths = ["tests"]
 pythonpath = ["."]
 
 [tool.radon]
